bitkeeper revision 1.170 (3e9c936fXyHEI0NKOWQkP9tHN4sbqw)
Many files:
  Finished virtualisation of x86 LDT. Xenolinux now exports this to applications (eg. for use by linuxthreads).

+++ xen/common/schedule.c
@@ -394,7 +394,7 @@ asmlinkage void schedule(void)
 
 #ifndef NDEBUG
     if (r_time < ctx_allow) {
-        printk("[%02d]: %lx\n", this_cpu, r_time);
+        printk("[%02d]: %lx\n", this_cpu, (unsigned long)r_time);
         dump_rqueue(&schedule_data[this_cpu].runqueue, "foo");
     }
 #endif
